requirements.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.7.0.72
Pillow==9.5.0
numpy==1.24.3
tqdm==4.65.0
scikit-learn==1.2.2
matplotlib==3.7.1